requirements.txt
Flask==2.3.3
twilio==9.0.1
python-dotenv==1.0.1
openai==1.30.1
langchain==0.1.16
langchain-openai==0.0.8
langchain-community==0.0.32
pypdf==5.7.0
